move ws request handling logic in websocket module
